requirements.txt
# Core PySpark
pyspark==3.5.0

# Data processing
pandas==2.1.3
numpy==1.24.3
pyarrow==14.0.1

# ML libraries
scikit-learn==1.3.2
mlflow==2.8.1

# Orchestration
apache-airflow==2.7.3
prefect==2.14.11

# Data quality
great-expectations==0.18.5

# Testing
pytest==7.4.3
pytest-spark==0.6.0
chispa==0.9.4

# Development
jupyter==1.0.0
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
